Restructure: markets
