+++ src/pages/markets/components/helper/_available-options.js
