+++ src/pages/markets/components/helper/_available-platforms.js
