+++ src/pages/markets/components/markets/_stock-indices.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Loadable from '@loadable/component'
-import { WhyTrade } from '../../_why-trade'
-import AvailableTrades from '../../_available-trades'
+import { WhyTrade } from '../sections/_why-trade'
+import AvailableTrades from '../helper/_available-trades'
 import stock_content from '../../static/content/_stock'
 import { stock_margin } from '../../static/content/_margin'
 import { stock_options } from '../../static/content/_digital-options'
@@ -10,7 +10,7 @@ import Margin from '../sub-markets/_margin'
 import DigitalOptions from '../sub-markets/_digital-options'
 import { localize, Localize } from 'components/localization'
 const SimpleSteps = Loadable(() => import('components/custom/_simple-steps'))
-const OtherMarkets = Loadable(() => import('../../_other-markets.js'))
+const OtherMarkets = Loadable(() => import('../sections/_other-markets.js'))
 
 const StockIndices = ({ simple_step_content }) => {
     simple_step_content[1].text = localize(
