+++ src/pages/markets/components/sections/_hero.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
-import Globe from './globe.mp4'
-import Globe2 from './globe.webm'
+import Globe from '../../static/video/globe.mp4'
+import Globe2 from '../../static/video/globe.webm'
 import { Container } from 'components/containers'
 import { Header } from 'components/elements'
 import { useLazyVideo } from 'components/hooks/lazy-video'
