+++ src/pages/markets/components/sub-markets/_multipliers.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import AvailablePlatforms from '../../_available-platforms.js'
-import { ContentWrapper, Descriptions, StyledText } from '../../_markets-style'
-import MarketInstruments from '../_market_instruments'
+import AvailablePlatforms from '../helper/_available-platforms.js'
+import { ContentWrapper, Descriptions, StyledText } from '../../static/style/_markets-style'
+import MarketInstruments from '../sections/_market_instruments'
 import { SectionContainer, Show } from 'components/containers'
 import { Text } from 'components/elements'
 import { localize } from 'components/localization'
