+++ src/pages/markets/index.js
@@ -4,7 +4,7 @@ import Commodities from './components/markets/_commodities.js'
 import Forex from './components/markets/_forex'
 import StockIndices from './components/markets/_stock-indices.js'
 import SyntheticIndices from './components/markets/_synthetic-indices.js'
-import { Hero } from './_hero'
+import { Hero } from './components/sections/_hero'
 import Signup, { Appearances } from 'components/custom/signup'
 import Layout from 'components/layout/layout'
 import { localize, Localize, WithIntl } from 'components/localization'
