+++ src/pages/markets/static/content/_margin.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Symbol from '../../components/_symbol'
+import Symbol from '../../components/helper/_symbol'
 import {
     CrashBoom,
     Energy,
