+++ src/pages/markets/static/style/_markets-style.js
