+++ src/pages/markets/sub-markets/_exotic-pairs.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Symbol from '../components/_symbol'
+import Symbol from '../components/helper/_symbol'
 import { exotic_pairs } from './_market-symbols'
 
 const ExoticPairs = () => {
