+++ src/pages/markets/sub-markets/_major-pairs.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Symbol from '../components/_symbol'
+import Symbol from '../components/helper/_symbol'
 import { major_pairs } from './_market-symbols'
 
 const MajorPairs = () => {
