+++ src/pages/markets/sub-markets/_step-indices.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Symbol from '../components/_symbol'
+import Symbol from '../components/helper/_symbol'
 import { STEPINDICES } from 'components/elements/symbols'
 import { localize } from 'components/localization'
