+++ src/pages/markets/sub-markets/_volatility-indices.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Symbol from '../components/_symbol'
+import Symbol from '../components/helper/_symbol'
 import { volatility_indices } from './_market-symbols'
 
 const VolatilityIndices = () => {
